requirements.txt
websocket-client:1.4.2